+++ GeneralWorks/Client/src/main/java/ServerUserInfo.java
@@ -6,7 +6,9 @@ public class ServerUserInfo {
 
 		public ServerUserInfo acceptLogin()
 		{
-			userinfo
+			//userinfo;
+            int a ;
+            return new ServerUserInfo();
 		}
 
 }
